Renamed files and folders for breaking_changes code
