+++ tools/diff-processor/breaking_changes/breaking_changes_test.go
@@ -1,4 +1,4 @@
-package rules
+package breaking_changes
 
 import (
 	"sort"
