+++ tools/diff-processor/breaking_changes/field_diff_test.go
@@ -1,4 +1,4 @@
-package rules
+package breaking_changes
 
 import (
 	"testing"
