+++ tools/diff-processor/breaking_changes/resource_diff.go
@@ -1,4 +1,4 @@
-package rules
+package breaking_changes
 
 import (
 	"fmt"
