+++ tools/diff-processor/breaking_changes/utility.go
@@ -1,4 +1,4 @@
-package rules
+package breaking_changes
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
