+++ tools/diff-processor/cmd/breaking_changes.go
@@ -11,7 +11,7 @@ import (
 	"sort"
 
 	"github.com/GoogleCloudPlatform/magic-modules/tools/diff-processor/diff"
-	"github.com/GoogleCloudPlatform/magic-modules/tools/diff-processor/rules"
+	"github.com/GoogleCloudPlatform/magic-modules/tools/diff-processor/breaking_changes"
 	"github.com/spf13/cobra"
 )
 
@@ -43,7 +43,7 @@ func newBreakingChangesCmd(rootOptions *rootOptions) *cobra.Command {
 }
 func (o *breakingChangesOptions) run() error {
 	schemaDiff := o.computeSchemaDiff()
-	breakingChanges := rules.ComputeBreakingChanges(schemaDiff)
+	breakingChanges := breaking_changes.ComputeBreakingChanges(schemaDiff)
 	sort.Slice(breakingChanges, func(i, j int) bool {
 		return breakingChanges[i].Message < breakingChanges[j].Message
 	})
